Backend modules/Complete backend/requirements.txt
# Core (tested & pinned)
numpy==1.24.3
opencv-python==4.8.1.78
mediapipe==0.10.11
scipy==1.10.1

# Voice / audio
praat-parselmouth==0.4.7
librosa==0.10.0
soundfile==0.12.1
pydub

# Transcription
openai-whisper

# NLP (Groq) – uses stable API, no version needed
groq
python-dotenv

# Speaker diarization (exact versions for compatibility)
pyannote.audio==3.1.1
huggingface-hub==0.27.0
torch
torchaudio

# Emotion detection (utilises HSEmotion)
torchvision
timm==0.9.2
hsemotion